+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Added
+- Allow copying a commit reference while replying without losing the draft state.
+
 ## [0.13.3] - 2026-07-18
 
 ### Added